Move _crefl_utils to modifiers subpackage
