+++ satpy/tests/test_crefl_utils.py
@@ -25,7 +25,7 @@ class TestCreflUtils(unittest.TestCase):
     def test_get_atm_variables_abi(self):
         """Test getting atmospheric variables for ABI."""
         import numpy as np
-        from satpy.composites.crefl_utils import get_atm_variables_abi
+        from satpy.modifiers._crefl_utils import get_atm_variables_abi
         sphalb, rhoray, TtotraytH2O, tOG = get_atm_variables_abi(0.17690244, 6.123234e-17, 530.61332168, 405.,
                                                                  21.71342113, 77.14385758, 56.214566960,
                                                                  0.0043149700000000004, 0.0037296,
